requirements.txt
Django>2.2,<3
django-nose==1.4.5
djangorestframework>3.9.1
mysqlclient>1.3.13
nose==1.3.7
django-cors-headers==2.1.0
django-model-utils>=3.1.2
Sphinx>2
sphinx-py3doc-enhanced-theme>=2.4.0
recommonmark==0.5.0
django-mysql
ipython
